+++ custom_components/stremio_stream_bridge/const.py
@@ -25,9 +25,17 @@ CONF_SUBTITLE_LANGUAGES = "subtitle_languages"
 CONF_SUBTITLE_CONVERT_VTT = "subtitle_convert_vtt"
 CONF_SUBTITLE_BASE_URL = "subtitle_base_url"
 
+DEFAULT_STREAMING_SERVER_URL = "http://192.168.1.145:11470"
 DEFAULT_CINEMETA_MANIFEST = "https://v3-cinemeta.strem.io/manifest.json"
 DEFAULT_TORRENTIO_MANIFEST = "https://torrentio.strem.fun/manifest.json"
 DEFAULT_OPENSUBTITLES_MANIFEST = "https://opensubtitles-v3.strem.io/manifest.json"
+DEFAULT_LATIN_MANIFEST = (
+    "https://torrentio.strem.fun/"
+    "providers=yts,eztv,rarbg,1337x,thepiratebay,kickasstorrents,torrentgalaxy,"
+    "magnetdl,horriblesubs,nyaasi,rutracker,mejortorrent,cinecalidad|sort=size|"
+    "language=spanish,latino|qualityfilter=brremux,hdrall,dolbyvision,4k/manifest.json"
+)
+DEFAULT_SPORTS_MANIFEST = "https://stremverse1.alwaysdata.net/manifest.json"
 DEFAULT_PREFERRED_QUALITY = "1080p"
 DEFAULT_MAX_SIZE_GB = 12.0
 DEFAULT_EXCLUDE_KEYWORDS = "CAM, HDCAM, TS, TELECINE, SCREENER"